Enum smartnoise_validator::base::Jagged

pub enum Jagged {
    Bool(Vec<Vec<bool>>),
    Int(Vec<Vec<Integer>>),
    Float(Vec<Vec<Float>>),
    Str(Vec<Vec<String>>),
}

The universal jagged array representation.

Typically used to store categorically clamped values. In practice, use is limited to public categories over multiple columns, and the upper triangular covariance matrix

Jagged has a one-to-one mapping to a protobuf Vector2DJagged.

Variants

Bool(Vec<Vec<bool>>)

Int(Vec<Vec<Integer>>)

Float(Vec<Vec<Float>>)

Str(Vec<Vec<String>>)

Implementations

impl Jagged

pub fn float(&self) -> Result<Vec<Vec<Float>>>

Retrieve the f64 jagged matrix, assuming the data type of the jagged matrix is f64, and assuming all columns are defined

pub fn int(&self) -> Result<Vec<Vec<Integer>>>

Retrieve the i64 jagged matrix, assuming the data type of the jagged matrix is i64

pub fn string(&self) -> Result<Vec<Vec<String>>>

Retrieve the String jagged matrix, assuming the data type of the jagged matrix is String

pub fn bool(&self) -> Result<Vec<Vec<bool>>>

Retrieve the bool jagged matrix, assuming the data type of the jagged matrix is bool

pub fn num_columns(&self) -> i64

pub fn num_records(&self) -> Vec<i64>

pub fn deduplicate(&self) -> Result<Jagged>

pub fn standardize(self, num_columns: i64) -> Result<Jagged>

pub fn data_type(&self) -> DataType

pub fn to_index_keys(&self) -> Result<Vec<Vec<IndexKey>>>