Struct smartnoise_validator::base::ArrayProperties

pub struct ArrayProperties {
    pub num_records: Option<i64>,
    pub num_columns: Option<i64>,
    pub nullity: bool,
    pub releasable: bool,
    pub c_stability: u32,
    pub aggregator: Option<AggregatorProperties>,
    pub nature: Option<Nature>,
    pub data_type: DataType,
    pub dataset_id: Option<i64>,
    pub node_id: i64,
    pub is_not_empty: bool,
    pub dimensionality: Option<i64>,
    pub group_id: Vec<GroupId>,
    pub naturally_ordered: bool,
    pub sample_proportion: Option<f64>,
}

Derived properties for the universal ArrayND.

The ArrayNDProperties has a one-to-one mapping to a protobuf ArrayNDProperties.

Fields
num_records: Option<i64>

Defined if the number of records is known statically (set by the resize component)

num_columns: Option<i64>nullity: bool

true if the data may contain null values

releasable: bool

set to true by the mechanisms. Acts as a filter on the values in the release

c_stability: u32

amplification of privacy usage by unstable data transformations, or possibility of duplicated records

aggregator: Option<AggregatorProperties>

set when data is aggregated, used to help compute sensitivity from the mechanisms

nature: Option<Nature>

either min/max or categories

data_type: DataType

f64, i64, bool, String

dataset_id: Option<i64>

index of last Materialize or Filter node, where dataset was created used to determine if arrays are conformable even when N is not known

node_id: i64

node index of the value

is_not_empty: bool

true if the number of rows is known to not be length zero

dimensionality: Option<i64>

number of axes in the array

group_id: Vec<GroupId>

used for tracking subpartitions

naturally_ordered: bool

used to determine if order of rows has changed

sample_proportion: Option<f64>

proportion of original data sampled

Implementations

impl ArrayProperties

pub fn lower(&self) -> Result<Array>

pub fn upper(&self) -> Result<Array>

pub fn lower_float_option(&self) -> Result<Vec<Option<Float>>>

pub fn lower_float(&self) -> Result<Vec<Float>>

pub fn upper_float_option(&self) -> Result<Vec<Option<Float>>>

pub fn upper_float(&self) -> Result<Vec<Float>>

pub fn lower_int_option(&self) -> Result<Vec<Option<Integer>>>

pub fn lower_int(&self) -> Result<Vec<Integer>>

pub fn upper_int_option(&self) -> Result<Vec<Option<Integer>>>

pub fn upper_int(&self) -> Result<Vec<Integer>>

pub fn categories(&self) -> Result<Jagged>

pub fn assert_non_null(&self) -> Result<()>

pub fn assert_is_not_empty(&self) -> Result<()>

pub fn assert_is_releasable(&self) -> Result<()>

pub fn num_columns(&self) -> Result<i64>

pub fn num_records(&self) -> Result<i64>

pub fn assert_is_not_aggregated(&self) -> Result<()>

pub fn assert_is_not_sampled(&self) -> Result<()>