Module smartnoise_validator::components::snapping_mechanism

Functions

accuracy_to_epsilon
Finds epsilon that will achieve desired accuracy and confidence requirements. Described in https://github.com/opendifferentialprivacy/smartnoise-core/blob/develop/whitepapers/mechanisms/snapping/snapping_implementation_notes.pdf

compute_precision
Finds the necessary precision for the snapping mechanism 118 bits required for LN -epsilon.log2().ceil() + 2 bits required for non-zero epsilon

epsilon_to_accuracy
Finds accuracy that is achievable given desired epsilon and confidence requirements. Described in https://github.com/opendifferentialprivacy/smartnoise-core/blob/develop/whitepapers/mechanisms/snapping/snapping_implementation_notes.pdf

get_smallest_greater_or_eq_power_of_two
Finds the smallest integer m such that 2^m is equal to or greater than x.

redefine_epsilon
Gets functional epsilon for Snapping mechanism such that privacy loss does not exceed the user's proposed budget. Described in https://github.com/opendifferentialprivacy/smartnoise-core/blob/develop/whitepapers/mechanisms/snapping/snapping_implementation_notes.pdf