Function smartnoise_validator::components::snapping_mechanism::epsilon_to_accuracy

pub fn epsilon_to_accuracy(
    alpha: f64, 
    epsilon: f64, 
    sensitivity: f64, 
    b: f64
) -> Result<f64>

Finds accuracy that is achievable given desired epsilon and confidence requirements. Described in https://github.com/opendp/smartnoise-core/blob/develop/whitepapers/mechanisms/snapping/snapping_implementation_notes.pdf

Arguments

alpha - Desired confidence level.

epsilon - Desired privacy guarantee.

sensitivity - l1 Sensitivity of function to which mechanism is being applied.

b - Upper bound on function value being privatized.

Returns

Accuracy of the Snapping mechanism.