Trait smartnoise_validator::components::Component

pub trait Component {
    fn propagate_property(
        &self, 
        privacy_definition: &Option<PrivacyDefinition>, 
        public_arguments: IndexMap<IndexKey, &Value>, 
        properties: NodeProperties, 
        _node_id: u32
    ) -> Result<Warnable<ValueProperties>>;
}

Universal Component trait

To be a component, a struct must represent an abstract computation, for which properties can be derived about the resulting data.

Required methods

fn propagate_property(
&self,
privacy_definition: &Option<PrivacyDefinition>,
public_arguments: IndexMap<IndexKey, &Value>,
properties: NodeProperties,
_node_id: u32
) -> Result<Warnable<ValueProperties>>

Given properties known about private arguments, and public arguments, derive properties about the resulting data.

A component must fail to propagate properties if requirements on the input properties are not met. For example, if a Component represents an abstract computation that requires prior knowledge of the number of records to be safe or function properly, the propagate_property implementation is expected to return an error state if the prior knowledge is not known.

For example, if a definition of privacy is used that is incompatible with the abstract computation, the propagate_property implementation is expected to return an error state.

Arguments

self - the protobuf object corresponding to the prost protobuf struct

privacy_definition - the definition of privacy under which the computation takes place

public_arguments - actual data values of arguments, typically either supplied literals or released values.

properties - derived properties of private input arguments

node_id - id of the node in the analysis graph (used to set dataset_id in the data loaders)

Returns

Derived properties on the data resulting from the abstract computation

Loading content...