Trait smartnoise_validator::components::Named

pub trait Named {
    fn get_names(
        &self, 
        public_arguments: IndexMap<IndexKey, &Value>, 
        argument_variables: IndexMap<IndexKey, Vec<IndexKey>>, 
        release: Option<&Value>
    ) -> Result<Vec<IndexKey>>;
}

Named component trait

Named components involve variables and keep track of the human readable names for these variables and may modify these variables names.

Required methods

fn get_names(
&self,
public_arguments: IndexMap<IndexKey, &Value>,
argument_variables: IndexMap<IndexKey, Vec<IndexKey>>,
release: Option<&Value>
) -> Result<Vec<IndexKey>>

Propagate the human readable names of the variables associated with this component

Loading content...