Trait smartnoise_validator::components::Report

pub trait Report {
    fn summarize(
        &self, 
        node_id: u32, 
        component: &Component, 
        public_arguments: IndexMap<IndexKey, &Value>, 
        properties: NodeProperties, 
        release: &Value, 
        variable_names: Option<&Vec<IndexKey>>
    ) -> Result<Option<Vec<JSONRelease>>>;
}

Report component trait

Reportable components correspond to a computation that a researcher may want a JSON summary for

Required methods

fn summarize(
&self,
node_id: u32,
component: &Component,
public_arguments: IndexMap<IndexKey, &Value>,
properties: NodeProperties,
release: &Value,
variable_names: Option<&Vec<IndexKey>>
) -> Result<Option<Vec<JSONRelease>>>

Summarize the relevant metadata around a computation in a readable, JSON-serializable format.

Loading content...